requirements.txt
Pillow>=10.0
pywin32>=306; sys_platform == "win32"
pyinstaller>=6.0
